+++ .replit
@@ -1,1 +1,1 @@
-run = "python main_2.py"
+run = "python main.py"
